move run(_:,arguments:) into SwiftTool
We'll need this to run the interpreter from SwiftBuildTool.
