+++ Sources/Commands/SwiftRunTool.swift
@@ -14,7 +14,6 @@ import Utility
 import PackageGraph
 import PackageModel
 
-import func POSIX.chdir
 import func POSIX.getcwd
 
 /// An enumeration of the errors that can be generated by the run tool.
@@ -165,18 +164,6 @@ public class SwiftRunTool: SwiftTool<RunToolOptions> {
             
             return rootExecutables[0]
         }
-    }
-    
-    /// Executes the executable at the specified path.
-    private func run(_ excutablePath: AbsolutePath, arguments: [String]) throws {
-        // Make sure we are running from the original working directory.
-        let cwd: AbsolutePath? = localFileSystem.currentWorkingDirectory
-        if cwd == nil || originalWorkingDirectory != cwd {
-            try POSIX.chdir(originalWorkingDirectory.asString)
-        }
-
-        let pathRelativeToWorkingDirectory = excutablePath.relative(to: originalWorkingDirectory)
-        try exec(path: excutablePath.asString, args: [pathRelativeToWorkingDirectory.asString] + arguments)
     }
 
     /// Determines if a path points to a valid swift file.
